Generic: add Nouveau and Nouveau-Legacy devices
Add experimental devices for Nouveau (GBM) and Nouveau-Legacy (X11)
to support older nVidia GPU's orphaned after removal of the 340.xx
vendor driver. This partially resurrects PR5917.

The oldest architecture that appears to be supported is the Tesla
(GeForce 8) series dating back to 2006. In testing, GeForce 6100
and 7300LE cards failed to run Xorg, although this needs further
investigation as users have reporting working installs with them
when using generic desktop distros like Slackware.

See [0] for info on VDPAU hardware decoding capabilities. Note that
H264 is not supported on all cards, and newer codecs like HEVC,VP9,
and AV1 are not supported at all. The GBM image has no support for
VDPAU decoding due to mesa state machine code depending on X11.

See [0] for info on cards that require firmware for VDPAU decoding
to work. It is not possible to embed the files as nVidia provides
them as part of the vendor driver and the license for the driver
forbids redistribution of driver file content. Thus we need to run
a download-extract script from userspace on first boot. This works
in the background if the HTPC device has an Ethernet connection to
the Internet. If the HTPC needs WiFi the nouveau-firmware service
will fail on boot until a connection has been configured.

[0] https://nouveau.freedesktop.org/VideoAcceleration.html

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>