Generic: add Nouveau-Legacy device
This partially resurrects PR5917 and supersedes PR9814 to add an
X11 based image to support older nVidia GPU's that were orphaned
after removal of the 340.xx vendor driver.

The oldest architecture that appears to be supported is the Tesla
(GeForce 8) series dating back to 2006. In testing, GeForce 6100
and 7300LE cards failed to start Xorg. This aligns support with the
previous vendor driver.

See [0] for info on VDPAU hardware decoding capabilities. Note that
H264 is not supported on all cards and HEVC, VP9, and AV1 are not
supported at all. VDPAU decoding in nouveau is dependent on the X11
state machine so while GBM images will run Kodi there is no VDPAU
support. Lakka will be able to add nouveau to the Generic image
while LibreELEC needs to define a separate Legacy image. It is not
possible to combine nouveau with the vendor driver in the existing
Legacy image due to overlaps.

See [0] for info on cards that require firmware for VDPAU decoding
to work. It is not possible to embed the files as nVidia provides
them as part of the vendor driver and the license for the driver
forbids redistribution of driver file content. Thus we need to run
a download-extract script from userspace on first boot. This works
in the background if the HTPC device has an Ethernet connection to
the Internet. If the HTPC needs WiFi the nouveau-firmware service
will fail on boot until a connection has been configured.

[0] https://nouveau.freedesktop.org/VideoAcceleration.html

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>